+++ src/candlewick/multibody/Visualizer.h
@@ -116,9 +116,13 @@ public:
 
   void processEvents();
 
-  bool shouldExit() const noexcept { return m_shouldExit; }
+  [[nodiscard]] bool shouldExit() const noexcept { return m_shouldExit; }
 
   void takeScreenshot(std::string_view filename);
+
+  void startRecording(std::string_view filename);
+
+  void stopRecording();
 
   /// \brief Clear objects
   void clean() override {